+++ packages/@expo/metro-runtime/src/location/Location.native.ts
@@ -20,8 +20,13 @@ class Location {
       href
     );
 
-    url.username = '';
-
+    try {
+      url.username = '';
+    } catch {
+      throw new Error(
+        'Attempting to use the window.location polyfill before the URL built-in has been polyfilled.'
+      );
+    }
     url.password = '';
     Object.defineProperties(this, {
       hash: {